Redesign the lua callback documentation
Father Linden, I have been brutally, some may even say savagely, attacked in the discord for not being able to understand the callback documentation. I have edited it for a simpler design with clear and specific use examples similar to how the JavaScript documentation for callbacks is. I hope you approve this, and the server documentation changes that will be coming in a bit, so I and similar smooth brain individuals have a better understanding.

The href in the client to server usage example may not be perfect I apologize in advance.